updated teckit to latest source from svn export teckit/TRUNK
git-svn-id: http://scripts.sil.org/svn-public/xetex/TRUNK@195 9f0520aa-a7ea-0310-80d7-d2976a5bef54

+++ TeX/libs/teckit/source/Compiler.h
@@ -332,7 +332,7 @@ protected:
 };
 
 struct CharName {
-	unsigned long	usv;
+	unsigned int	usv;
 	const char*		name;
 };
 extern CharName	gUnicodeNames[];
